[spotify] Improve performance of handling data
When Spotify returns a 204. It means it doesn't return any data. This can happen after a certain time of inactivity. When this is returned it doesn't make sense to update the states or poll for devices/playlists either.
Also let playlist refresh only once per hour or when an explicit refresh of the channel is given will it update on the next poll.

Signed-off-by: Hilbrand Bouwkamp <hilbrand@h72.nl>